Find select2 when it's added asynchronously.
The scenario: press a button that (perhaps via AJAX) causes a select2
with a label to appear asynchronously. This test would fail:

    click_button('Make a Select2')
    select2('My Choice', from: 'The Newly Created Select2')

because the label isn't present immediately, and `first("label", ...)`
doesn't wait for the label to appear. `find` does.

This change will cause an error to be raised if there's more than one label
with the given text. The solution here is to limit the scope using
`within`, and/or use a selector instead of text to find the select2.