More robust if label and select2 aren't adjacent.
The previous code expected the select2 to be inside the innermost div
containing the label. That won't necessarily hold for all scenarios (for
example, the "Inline Labels" markup for Zurb Foundation). So now we use
the label to find the element that it gives focus to, and then search up
the DOM to find the select2-container.