Handle both single and multiple selects w/ Poltergeist.
In the single-select scenario, we clicked the "choice" element (to open
the dropdown), then clicked it again during the "values" loop. This
probably works on some Capybara drivers, but Poltergeist noticed that
the second time we tried to click, the choice element was obscured by
select2's dropdown mask.

There's no reason to click the choice element twice in the single-select
UI. But we do need to click it once for each value in the
multiple-select UI. So now we just check which UI is in use, and act
appropriately.

BTW, we do the single-select click *before* we enter the search term. I
suspect search doesn't work in the multiple-select UI, but it didn't
work before this change, either.